pyproject.toml
[tool.poetry]
name = "iroh_testlab"
version = "0.1.0"
description = ""
authors = ["Iroh <iroh@mail.com>"]

[tool.poetry.dependencies]
python = "^3.9.9"
aiohttp = "3.8.5"
asyncpg = "^0.28.0"
pyyaml =  "6.0"
pydantic = "^1.10.4"
loguru = "0.7"
pillow = "10.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
